add hostname as a label in MetaObject.
When we do 'kubectl describe disk', we get all the disks for the cluster.
We can not relate which disk is on which node. The node details in
the MetaObject will be helpful.

Also when we are restarting the node-disk-manager, we are deleting the
disk resources from etcd if the resource is detached in between stop and start.
It may be a temporary outage so we should not delete it, instead we should have
a `status` field for the disk resources which will signify that the disk
resource is Active or Inactive and if user wants to remove the disk completely,
we have to execute `purge` command(TODO), which will clean up these kind of
resources from etcd.

Signed-off-by: Pawan <pawanprakash101@gmail.com>