populating the hostnames correctly for the Disk resources.
Pods will have hostname as pod name and some hash string
(e.g. node-disk-manager-29qw8) which changes with every pod
restart.

We are using system's hostname for populating the
disk resource in etcd. Disk resource should not use the hostname
of the pod for "kubernetes.io/hostname", it should be node's hostname,
which will tell us where the disk is attached to.

Here, we are passing the hostname as an environment to the Node-Disk-Manager
daemonset, which it will use to populate the disk resource in etcd.

Signed-off-by: Pawan <pawanprakash101@gmail.com>