creating a single yaml for the deployment of NDM
It is cumbersome to apply more than one yaml to deploy a solution.
For NDM, we are applying crd yaml and deployment yaml for deployment.
Writing a complete yaml, which will do everyting like apply the
CRD yaml, create a service account, apply the deployment yaml would
be useful for users.

Signed-off-by: Pawan <pawanprakash101@gmail.com>