Include support for creating sparse file
For testing with nodes that have no additional disks, sparse disks provide a
mechanism to create a sparse file and use as a disk resource.

The NDM sparse file creation can be configured via the following
environment variables:
- SPARSE_FILE_DIR : A valid directory where sparse files need to be created
- SPARSE_FILE_SIZE : Size of each sparse file in byte. Defaults to 1GB
- SPRASE_FILE_COUNT: Number of files to be created. Defaults to 1

To differentiate between sparse disks and regular disk, a disk-type label
as been introduced. For sparse disks, this disk-type will be `sparse`, where
as for regular disks it will be set as `disk`.

Like disk resources, the sparse disk resource's state will be marked as
`Uknown` when NDM is shutdown and is made `Active` when the NDM is restarted
on the same node.

Note:
- If a sparse disk already exists in the directory, it will be reused as is.
- The naming convention for sparse disk is <user specified dir>/<count>-ndm-sparse.img
  count starts with 0 to n-1, where n is the count specified by user as
  SPARSE_FILE_COUNT

Signed-off-by: kmova <kiran.mova@openebs.io>