Fix sparse disks state mismatch
ndm makes the state of sparse file related cr as inactive if any
mismatch occur. One case is like this if user is using openEBS
volumes then wrapper iscsi disk is created and vendor of that disk
is openEBS which we are filtering out. When user delete that volume
one disk detach event comes and as that disk is not present in etcd
it is considered as a mismatch. Then ndm will rescan system and update
state of the crs. As rescan is dependent on udev only it will not
identify sparse files and make the state of the cr as inactive.

This PR contains the code which will check for sparse files if it is
sparse file related cr and it will make the status of the cr inactive
if sparse file is not present.

Changes committed:
modified:   cmd/controller/diskstore.go
modified:   cmd/controller/sparsefilegenerator.go
modified:   cmd/controller/sparsefilegenerator_test.go

Signed-off-by: Shovan Maity <shovan.maity@mayadata.io>