(feat): add support for custom disk (#192)
NDM works by detecting the devices attached to a node and involves marking
the devices (disk CRs) as active or unknown etc., depending on the node status. 
However, there are some cases which came up where users wanted to use a partitioned 
device rather than a full disk. 

To support such cases, user would like to manually a provision a Disk CR
that should not be managed by NDM. 

This PR add the support, which allow a disk CR to be unmanaged by adding the following label: 
 `ndm.io/managed=false`

If the label is not present, the value is assumed to be true. 

Signed-off-by: Shovan Maity <shovan.cse91@gmail.com>