fix(install): changed pull policy and upgrade strategy (#200)
Changed `ImagePullPolicy` to `IfNotPresent`. Earlier, during integration testing the old image from docker-hub was pulled which overwrite the newly built images. This resulted in integration tests being performed on the old image and also the old image was re-tagged and pushed. With the change, only the newly built image will be used.

The update strategy of Daemonset is now set to RollingUpdate, so that the new image will be applied without manual deletion.

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>