[US4295] feat(diskCR): detecting change/delete udev events
Add/delete events for partitions will be now monitored and the changes
corresponding will be updated in the diskCR. When a partition is deleted
the last partition data will be removed from the diskCR, instead of
removing the exact partition as a unique id to detect partitions is yet
to be implemented. changes in disk/devices (add/delete) will be now
monitored through udev events.

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>