feat (BDD) : NDM integration testing
1) Added YAMLs for integration testing
2) Udev functions for simulating devices. The udev functions could be
used to trigger fake udev events so that ndm detects the events and
updates the CRs. functions will be required to create a fake disk,
trigger add/remove/change events
3) sanity tests will be done in a single node minikube cluster and
functions for interfacing with the cluster is included.The cluster
start and stop is not included as this will be done by the VM.
Status of cluster, and generation of kube-config files will be checked
4) package to query kube-API server and retrieve results in a predefined
format is included. The various resources like configmap, CRD etc can
also be created using this package. These resources will be created
using the YAMLs
5) The basic tests include:
        a) running NDM and checking for pod
        b) checking for sparse disks
        c) check for external disk already attached
        d) check for external disk attached at runtime

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>