fix(cleaner): start cleanup jobs only on active BDs
This fix makes sure that cleanup jobs are started only on Active
BDs. Also, if the BD goes into another state, during cleanup, the
job will be cancelled and BD state will remain unchanged. This ensures that stale jobs are not present in the cluster, when corresponding BDs have been removed.

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>