feat(apis, controller, manager): include node attributes to disk/BD (#296)
Currently only hostname was included as part of the device information. This has been refactored to include nodename and hostname along with provision for multiple other fields which can be used for geolocating a node and its associated disks and BDs. Also, there are use cases where nodeName and hostName are different. 

This PR has made the following changes:
- add a nodeAttribute section to the BD YAML
- replace hostname with node attributes in the DiskInfo. 
  Node attributes will contain hostname along with other 
  attributes of the node on which the daemon is running. 
  The attribute can be nodename, zone, failure-domain etc
- refactor node attributes (hostname and nodename) across 
  disks, BDs and sparse devices.
- update the sample BD yaml with node attribute

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>