refactor(build): for multiple arch compilation using gox (#308)
This PR refactors the build scripts ( .travis.yaml and Makefile) to allow
for compiling on different platforms like amd64 and arm64 without
having to make changes to the build scripts.

Related to:

[openebs/openebs#1295](https://github.com/openebs/openebs/issues/1295)
[#306](https://github.com/openebs/node-disk-manager/pull/306)


The following changes have been made:


* shellcheck is triggered as a seperate step from .travis.yaml, as it is only available for amd64
* updated the Makefile to initialize ARCH and BASEIMAGE depending on the platform where builds are initiated.
* updated the build to make use of gox to allow for cross compilation
* made some cosmetic changes to Makefile for better readability.
* updated the README with dependency on SeaChest libraries.

Attempted to automate the building of arm64 builds via cross compiling
in travis. However, the dependencies on libudev and seachest are non-trivial
to resolve. Will re-attempt to setup automated builds via QEMU based approach
in the upcoming PRs.

Note: Verified the changes on ARM by manually running the - make in a ARM node.

Signed-off-by: kmova kiran.mova@mayadata.io