feat(build, version): add version and commit to builds (#338)
Version and GitCommit is now added as part of NDM build. This will enable the binaries to log commit from which the build was generated and also the version of NDM.
By default ci will be the version used. 
For releases and RC builds the tags will be used as version.


* add scripts for building with version and git commit
* add version tag and git commit while building the binaries.
* refactor pre run func()
* added version information to ndm operator

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>