fix(tools): set the default tag to latest for helper pod
The default helper pod which can be overridden via the
ENV was set to fixed older version. This can result in
default image to have some security vulnerabilities.

As part of the release process, the helper pod images will
be rebuilt allowing to pull in a image that has fixed
known security vulnerabilities.

Signed-off-by: kmova <kiran.mova@mayadata.io>