fix(script): fix changing the system core pattern
earlier NDM used to rewrite the system core pattern so that the core
dump is available in the persistent /var/openebs/sparse directory.
But this will cause all process to write core files to that directory.
Now NDM process will be launched from the openebs directory, so core
files will get automatically written to the $PWD. This will help to
remove changing the system core pattern.

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>