refact(seachest,probe): remove unused attributes from seachest probe
- attributes which are guaranteed to be filled by other probes are removed
from seachest. seachest-probe will be used to fetch details that are
exclusively given by seachest library.
- some fields like RotationRate, Endurance etc has been removed as they
are still not yet supported by the BlockDevice resource.
- seachest wont't be used to fill in storage/capacity related fields,
because the library will always report the capacity/blocksize of the main
disk even if the query is made for the partition. This will result in wrong
information being filled in the blockdevice resource. The sysfs probe will be
used to get all capacity related information from the sysfs filesystem

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>