fix(controller): fix BDC controller accidentally updating selector
BDC controller used a pointer reference of the selector. This
has been changed to copy the selector and use it. This is done so that
an update on the BDC object does not cause the selector to get
updated with a different value other than the one which was applied by
user.

Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>