fix(cleaner); fix cleaning partitions before wiping main disk (#445)
when an application creates partitions on a claimed disk, first the filesystem signature
in the partitions need to be wiped and then the partition table. Otherwise, recreating
the partition table will result in the system picking up the filesystem signature from the
previous partition. Also partprobe command is invoked so as to re-read the partition table,
after wiping the disk


Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>