fix(probe): add unknown drive type (#523)
add unknown drive type to blockdevices. This will be used instead of the empty string. While adding the crd validation we came across this case which failed validation due to code setting to blank. Instead of blank, if we provide Unknown it will be better in terms of readability and validation.


Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>