refactor(mount): pass file path as parameter in DeviceBasicMountInfo() (#539)
the `Identifier.DeviceBasicMountInfo()` fetches mount attributes of a
device by reading the mounts file. The path of this file is currently
fixed. Pass the file path as a parameter to the function. This will
provide more flexibility in the code and also ease writing unit tests
for code that uses this function. The default path is exported as a
constant, and can still be used.

Signed-off-by: Aditya Jain <aditya.jainadityajain.jain@gmail.com>