feat(build): add ability to push to multiple registries from github actions (#544)
Currently the multiarch images can be pushed only to a single registry without rebuilding.  Therefore the images are by default pushed to docker hub only. This is not ideal and images should be pushed to multiple registries. This PR fixes this issue by making use of the official docker [build-push-action](https://github.com/docker/build-push-action).

Refactored the Github actions to:
- make use of docker build-push-action
- push images to quay and docker


Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>