feat(pkg, libmount): add new package to parse mount files
libmount is a pure go port of the util-linux/libmount library. while
most of the original functionalities are missing in this go
implementation, it currently supports parsing mounts file in procfs and
generating diffs between two parsed mounts files. while the current
functionalities are enough to solve some of the issues faced in #595,
this is a wip and remaining functions maybe added in future.

Signed-off-by: Aditya Jain <aditya.jainadityajain.jain@gmail.com>